migrate UIApplicationClient
