+++ WegliKit/Sources/UIApplicationClient/Mock.swift
@@ -1,0 +1,9 @@
+import ComposableArchitecture
+import UIKit.UIApplication
+
+public extension UIApplicationClient {
+  static let noop = Self(
+      open: { _, _ in .none },
+      openSettingsURLString: { "" }
+  )
+}
